LayoutTests/fast/viewport/viewport-69-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument key "xinitial-scale" not recognized and ignored.
viewport size 100x110 scale 3.200000 with limits [3.200000, 5.000000]
